+++ packages/std-client/src/setup/setupMUDNetwork.ts
@@ -5,13 +5,16 @@ import {
   createTxQueue,
   createSyncWorker,
   createEncoder,
-  NetworkComponentUpdate,
   createSystemExecutor,
   NetworkConfig,
   SyncWorkerConfig,
   isNetworkComponentUpdateEvent,
+  NetworkEvent,
+  ack,
+  Ack,
+  InputType,
 } from "@latticexyz/network";
-import { BehaviorSubject, bufferTime, filter, Observable, Subject } from "rxjs";
+import { BehaviorSubject, concatMap, filter, from, map, Observable, Subject, timer } from "rxjs";
 import {
   Component,
   Components,
@@ -25,7 +28,7 @@ import {
   World,
 } from "@latticexyz/recs";
 import { computed, IComputedValue } from "mobx";
-import { keccak256, stretch, toEthAddress } from "@latticexyz/utils";
+import { keccak256, toEthAddress } from "@latticexyz/utils";
 import ComponentAbi from "@latticexyz/solecs/abi/Component.json";
 import { Contract, ContractInterface, Signer } from "ethers";
 import { Component as SolecsComponent } from "@latticexyz/solecs";
@@ -35,10 +38,20 @@ import { abi as WorldAbi } from "@latticexyz/solecs/abi/World.json";
 import { defineStringComponent } from "../components";
 
 export type SetupContractConfig = NetworkConfig &
-  Omit<SyncWorkerConfig, "worldContract" | "mappings"> & { worldAddress: string; devMode?: boolean };
+  Omit<SyncWorkerConfig, "worldContract" | "mappings"> & {
+    worldAddress: string;
+    devMode?: boolean;
+  };
+
+export type ContractComponent = Component<Schema, { contractId: string }>;
 
 export type ContractComponents = {
-  [key: string]: Component<Schema, { contractId: string }>;
+  [key: string]: ContractComponent;
+};
+
+export type NetworkComponents<C extends ContractComponents> = C & {
+  SystemsRegistry: Component<{ value: Type.String }>;
+  ComponentsRegistry: Component<{ value: Type.String }>;
 };
 
 export async function setupMUDNetwork<C extends ContractComponents, SystemTypes extends { [key: string]: Contract }>(
@@ -58,16 +71,21 @@ export async function setupMUDNetwork<C extends ContractComponents, SystemTypes 
     metadata: { contractId: "world.component.components" },
   });
 
-  components = {
-    ...components,
-    SystemsRegistry,
-    ComponentsRegistry,
-  };
+  (components as NetworkComponents<C>).SystemsRegistry = SystemsRegistry;
+  (components as NetworkComponents<C>).ComponentsRegistry = ComponentsRegistry;
 
+  // Mapping from component contract id to key in components object
   const mappings: Mappings<C> = {};
-  for (const key of Object.keys(components)) {
-    const { contractId } = components[key].metadata;
+
+  // Function to register new components in mappings object
+  function registerComponent(key: string, component: ContractComponent) {
+    const { contractId } = component.metadata;
     mappings[keccak256(contractId)] = key;
+  }
+
+  // Register initial components in mappings object
+  for (const key of Object.keys(components)) {
+    registerComponent(key, components[key]);
   }
 
   const network = await createNetwork(networkConfig);
@@ -90,33 +108,38 @@ export async function setupMUDNetwork<C extends ContractComponents, SystemTypes 
   });
   world.registerDisposer(disposeTxQueue);
 
-  const systems = createSystemExecutor<SystemTypes>(world, network, SystemsRegistry, SystemAbis, gasPriceInput$, {
-    devMode: networkConfig.devMode,
-  });
+  const { systems, registerSystem } = createSystemExecutor<SystemTypes>(
+    world,
+    network,
+    SystemsRegistry,
+    SystemAbis,
+    gasPriceInput$,
+    {
+      devMode: networkConfig.devMode,
+    }
+  );
 
   // Create sync worker
-  const { ecsEvent$, config$, dispose } = createSyncWorker<C>();
+  const ack$ = new Subject<Ack>();
+  const { ecsEvents$, input$, dispose } = createSyncWorker<C>(ack$);
   world.registerDisposer(dispose);
   function startSync() {
-    config$.next({
-      provider: networkConfig.provider,
-      worldContract: contractsConfig.World,
-      initialBlockNumber: networkConfig.initialBlockNumber ?? 0,
-      chainId: networkConfig.chainId,
-      disableCache: networkConfig.devMode, // Disable cache on local networks (hardhat / anvil)
-      snapshotServiceUrl: networkConfig.snapshotServiceUrl,
-      streamServiceUrl: networkConfig.streamServiceUrl,
+    input$.next({
+      type: InputType.Config,
+      data: {
+        ...networkConfig,
+        worldContract: contractsConfig.World,
+        initialBlockNumber: networkConfig.initialBlockNumber ?? 0,
+        disableCache: networkConfig.devMode, // Disable cache on local networks (hardhat / anvil)
+      },
     });
   }
 
-  const { txReduced$ } = applyNetworkUpdates(
-    world,
-    components,
-    ecsEvent$.pipe(filter(isNetworkComponentUpdateEvent)),
-    mappings
-  );
+  const { txReduced$ } = applyNetworkUpdates(world, components, ecsEvents$, mappings, ack$);
 
-  const encoders = createEncoders(world, ComponentsRegistry, signerOrProvider);
+  const encoders = networkConfig.encoders
+    ? createEncoders(world, ComponentsRegistry, signerOrProvider)
+    : new Promise((resolve) => resolve({}));
 
   return {
     txQueue,
@@ -126,10 +149,10 @@ export async function setupMUDNetwork<C extends ContractComponents, SystemTypes 
     startSync,
     systems,
     gasPriceInput$,
-    ecsEvent$,
+    ecsEvent$: ecsEvents$.pipe(concatMap((updates) => from(updates))),
     mappings,
-    SystemsRegistry,
-    ComponentsRegistry,
+    registerComponent,
+    registerSystem,
   };
 }
 
@@ -169,39 +192,49 @@ async function createEncoders(
 function applyNetworkUpdates<C extends Components>(
   world: World,
   components: C,
-  ecsEvent$: Observable<NetworkComponentUpdate<C>>,
-  mappings: Mappings<C>
+  ecsEvents$: Observable<NetworkEvent<C>[]>,
+  mappings: Mappings<C>,
+  ack$: Subject<Ack>
 ) {
   const txReduced$ = new Subject<string>();
 
-  const ecsEventSub = ecsEvent$
+  // Send "ack" to tell the sync worker we're ready to receive events while not processing
+  let processing = false;
+  const ackSub = timer(0, 100)
     .pipe(
-      // We throttle the client side event processing to 1000 events every 16ms, so 62.500 events per second.
-      // This means if the chain were to emit more than 62.500 events per second, the client would not keep up.
-      // The only time we get close to this number is when initializing from a snapshot/cache.
-      bufferTime(16, null, 1000),
-      filter((updates) => updates.length > 0),
-      stretch(16)
+      filter(() => !processing),
+      map(() => ack)
     )
-    .subscribe((updates) => {
-      // Running this in a mobx action would result in only one system update per frame (should increase performance)
-      // but it currently breaks defineUpdateAction (https://linear.app/latticexyz/issue/LAT-594/defineupdatequery-does-not-work-when-running-multiple-component)
-      for (const update of updates) {
-        const entityIndex = world.entityToIndex.get(update.entity) ?? world.registerEntity({ id: update.entity });
-        const componentKey = mappings[update.component];
-        if (!componentKey) return console.warn("Unknown component:", update);
+    .subscribe(ack$);
 
-        if (update.value === undefined) {
-          // undefined value means component removed
-          removeComponent(components[componentKey] as Component<Schema>, entityIndex);
-        } else {
-          setComponent(components[componentKey] as Component<Schema>, entityIndex, update.value);
-        }
+  const delayQueueSub = ecsEvents$.subscribe((updates) => {
+    processing = true;
+    for (const update of updates) {
+      if (!isNetworkComponentUpdateEvent<C>(update)) continue;
+      if (update.lastEventInTx) txReduced$.next(update.txHash);
 
-        if (update.lastEventInTx) txReduced$.next(update.txHash);
+      const entityIndex = world.entityToIndex.get(update.entity) ?? world.registerEntity({ id: update.entity });
+      const componentKey = mappings[update.component];
+      if (!componentKey) {
+        console.warn("Unknown component:", update);
+        continue;
       }
-    });
 
-  world.registerDisposer(() => ecsEventSub?.unsubscribe());
+      if (update.value === undefined) {
+        // undefined value means component removed
+        removeComponent(components[componentKey] as Component<Schema>, entityIndex);
+      } else {
+        setComponent(components[componentKey] as Component<Schema>, entityIndex, update.value);
+      }
+    }
+    // Send "ack" after every processed batch of events to process faster than ever 100ms
+    ack$.next(ack);
+    processing = false;
+  });
+
+  world.registerDisposer(() => {
+    delayQueueSub?.unsubscribe();
+    ackSub?.unsubscribe();
+  });
   return { txReduced$: txReduced$.asObservable() };
 }